requirements.txt
python-telegram-bot==13.11
Flask==2.1.1
flask-sqlalchemy